Correct some accesibility oversights
1. Restore LFO buttons and Modulation buttons
2. Refactor the tab order stuff a bit
3. Miscellany

Addresses #4616
