+++ src/surge-xt/gui/AccessibleHelpers.h
@@ -29,6 +29,65 @@ namespace Surge
 namespace Widgets
 {
 
+struct GroupTagTraverser : public juce::ComponentTraverser
+{
+    /*
+     * This is a crude implementation which just sorts all components
+     */
+    juce::Component *on{nullptr};
+    GroupTagTraverser(juce::Component *c) : on(c) {}
+    juce::Component *getDefaultComponent(juce::Component *parentComponent) override
+    {
+        return nullptr;
+    }
+    juce::Component *getNextComponent(juce::Component *current) override { return nullptr; }
+    juce::Component *getPreviousComponent(juce::Component *current) override { return nullptr; }
+    std::vector<juce::Component *> getAllComponents(juce::Component *parentComponent) override
+    {
+        std::vector<juce::Component *> res;
+        for (auto c : on->getChildren())
+            res.push_back(c);
+        std::sort(res.begin(), res.end(), [this](auto a, auto b) { return lessThan(a, b); });
+        return res;
+    }
+
+    bool lessThan(const juce::Component *a, const juce::Component *b)
+    {
+        int acg = -1, bcg = -1;
+        auto ap = a->getProperties().getVarPointer("ControlGroup");
+        auto bp = b->getProperties().getVarPointer("ControlGroup");
+        if (ap)
+            acg = *ap;
+        if (bp)
+            bcg = *bp;
+
+        if (acg != bcg)
+            return acg < bcg;
+
+        auto at = dynamic_cast<const Surge::GUI::IComponentTagValue *>(a);
+        auto bt = dynamic_cast<const Surge::GUI::IComponentTagValue *>(b);
+
+        if (at && bt)
+            return at->getTag() < bt->getTag();
+
+        if (at && !bt)
+            return false;
+
+        if (!at && bt)
+            return true;
+
+#if SURGE_JUCE_ACCESSIBLE
+        auto cd = a->getDescription().compare(b->getDescription());
+        if (cd < 0)
+            return true;
+        if (cd > 0)
+            return false;
+#endif
+
+        // so what the hell else to do?
+        return a < b;
+    }
+};
 template <typename T> struct DiscreteAHRange
 {
     static int iMaxV(T *t) { return t->iMax; }
